borrowing cast pre perzistentnu vrstvu

+++ src/main/java/umb/fpv/ki/bookLibrary/BorrowingRepository.java
@@ -2,6 +2,6 @@ package umb.fpv.ki.bookLibrary;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BorrowingRepository extends JpaRepository<Borrowing, Integer> {
+public interface BorrowingRepository extends JpaRepository<Borrowing, Long> {
 
 }
